+++ packages/config-plugins/src/Plugin.types.ts
@@ -68,6 +68,10 @@ export type Mod<Props = any> = (
 export interface ModConfig {
   android?: {
     /**
+     * Dangerously make a modification before any other android mods have been run.
+     */
+    dangerous?: Mod<unknown>;
+    /**
      * Modify the `android/app/src/main/AndroidManifest.xml` as JSON (parsed with [`xml2js`](https://www.npmjs.com/package/xml2js)).
      */
     manifest?: Mod<AndroidManifest>;
@@ -97,6 +101,10 @@ export interface ModConfig {
     gradleProperties?: Mod<Properties.PropertiesItem[]>;
   };
   ios?: {
+    /**
+     * Dangerously make a modification before any other android mods have been run.
+     */
+    dangerous?: Mod<unknown>;
     /**
      * Modify the `ios/<name>/Info.plist` as JSON (parsed with [`@expo/plist`](https://www.npmjs.com/package/@expo/plist)).
      */
