+++ packages/config-plugins/src/plugins/__tests__/mod-compiler-test.ts
