+++ packages/config-plugins/src/plugins/__tests__/withMod-test.ts
@@ -1,0 +1,37 @@
+import { ExportedConfig } from '../../Plugin.types';
+import { evalModsAsync } from '../mod-compiler';
+import { withMod } from '../withMod';
+
+describe(withMod, () => {
+  it('compiles mods', async () => {
+    // A basic plugin exported from an app.json
+    const exportedConfig: ExportedConfig = { name: 'app', slug: '', mods: null };
+
+    // Apply mod
+    let config = withMod<any>(exportedConfig, {
+      platform: 'android',
+      mod: 'custom',
+      action(props) {
+        // Capitalize app name
+        props.name = (props.name as string).toUpperCase();
+        return props;
+      },
+    });
+
+    // Compile plugins generically
+    config = await evalModsAsync(config, { projectRoot: '/' });
+
+    // Plugins should all be functions
+    expect(Object.values(config.mods.android).every(value => typeof value === 'function')).toBe(
+      true
+    );
+
+    delete config.mods;
+
+    // App config should have been modified
+    expect(config).toStrictEqual({
+      name: 'APP',
+      slug: '',
+    });
+  });
+});
